Rename Recommender to PureRecommender
