+++ baybe/recommenders/__init__.py
@@ -1,18 +1,23 @@
 """BayBE recommenders."""
 
-from baybe.recommenders.bayesian.sequential_greedy import SequentialGreedyRecommender
 from baybe.recommenders.meta.sequential import (
     SequentialMetaRecommender,
     StreamingSequentialMetaRecommender,
     TwoPhaseMetaRecommender,
 )
 from baybe.recommenders.naive import NaiveHybridSpaceRecommender
-from baybe.recommenders.nonpredictive.clustering import (
+from baybe.recommenders.pure.bayesian.sequential_greedy import (
+    SequentialGreedyRecommender,
+)
+from baybe.recommenders.pure.nonpredictive.clustering import (
     GaussianMixtureClusteringRecommender,
     KMeansClusteringRecommender,
     PAMClusteringRecommender,
 )
-from baybe.recommenders.nonpredictive.sampling import FPSRecommender, RandomRecommender
+from baybe.recommenders.pure.nonpredictive.sampling import (
+    FPSRecommender,
+    RandomRecommender,
+)
 
 __all__ = [
     "FPSRecommender",
