--- baybe/recommenders/bayesian/__init__.py
@@ -1,7 +1,0 @@
-"""Bayesian recommenders."""
-
-from baybe.recommenders.bayesian.sequential_greedy import SequentialGreedyRecommender
-
-__all__ = [
-    "SequentialGreedyRecommender",
-]
