+++ baybe/recommenders/pure/__init__.py
@@ -1,0 +1,19 @@
+"""Pure recommenders."""
+
+from baybe.recommenders.pure.bayesian import SequentialGreedyRecommender
+from baybe.recommenders.pure.nonpredictive import (
+    FPSRecommender,
+    GaussianMixtureClusteringRecommender,
+    KMeansClusteringRecommender,
+    PAMClusteringRecommender,
+    RandomRecommender,
+)
+
+__all__ = [
+    "SequentialGreedyRecommender",
+    "RandomRecommender",
+    "FPSRecommender",
+    "PAMClusteringRecommender",
+    "KMeansClusteringRecommender",
+    "GaussianMixtureClusteringRecommender",
+]
