+++ baybe/recommenders/pure/bayesian/__init__.py
@@ -1,0 +1,9 @@
+"""Bayesian recommenders."""
+
+from baybe.recommenders.pure.bayesian.sequential_greedy import (
+    SequentialGreedyRecommender,
+)
+
+__all__ = [
+    "SequentialGreedyRecommender",
+]
