+++ baybe/recommenders/pure/nonpredictive/__init__.py
@@ -1,11 +1,14 @@
 """Nonpredictive recommenders."""
 
-from baybe.recommenders.nonpredictive.clustering import (
+from baybe.recommenders.pure.nonpredictive.clustering import (
     GaussianMixtureClusteringRecommender,
     KMeansClusteringRecommender,
     PAMClusteringRecommender,
 )
-from baybe.recommenders.nonpredictive.sampling import FPSRecommender, RandomRecommender
+from baybe.recommenders.pure.nonpredictive.sampling import (
+    FPSRecommender,
+    RandomRecommender,
+)
 
 __all__ = [
     "FPSRecommender",
