+++ docs/userguide/campaigns.md
@@ -97,7 +97,7 @@ used is strongly discouraged.
 **Note:** While the above distinction is true in the general case, it may not be 
 relevant for all configured settings, for instance, when the used recommender 
 is not capable of joint optimization. Currently, the 
-[SequentialGreedyRecommender](baybe.recommenders.bayesian.sequential_greedy.SequentialGreedyRecommender)
+[SequentialGreedyRecommender](baybe.recommenders.pure.bayesian.sequential_greedy.SequentialGreedyRecommender)
 is the only recommender available that performs joint optimization.
 ```
