+++ docs/userguide/recommender.md
@@ -8,9 +8,9 @@ The set of available recommenders can be partitioned into the following subclass
 
 ## Bayesian recommenders
 
-The Bayesian recommenders in BayBE are built on the foundation of the [`BayesianRecommender`](baybe.recommenders.bayesian.base.BayesianRecommender) class, offering an array of possibilities with internal surrogate models and support for various acquisition functions.
+The Bayesian recommenders in BayBE are built on the foundation of the [`BayesianRecommender`](baybe.recommenders.pure.bayesian.base.BayesianRecommender) class, offering an array of possibilities with internal surrogate models and support for various acquisition functions.
 
-The [`SequentialGreedyRecommender`](baybe.recommenders.bayesian.sequential_greedy.SequentialGreedyRecommender) is a powerful recommender that leverages BoTorch optimization functions to perform sequential Greedy optimization. It can be applied for discrete, continuous and hybrid sarch spaces. It is an implementation of the BoTorch optimization functions for discrete, continuous and mixed spaces. 
+The [`SequentialGreedyRecommender`](baybe.recommenders.pure.bayesian.sequential_greedy.SequentialGreedyRecommender) is a powerful recommender that leverages BoTorch optimization functions to perform sequential Greedy optimization. It can be applied for discrete, continuous and hybrid sarch spaces. It is an implementation of the BoTorch optimization functions for discrete, continuous and mixed spaces. 
 
 It is important to note that this recommender performs a brute-force search when applied in hybrid search spaces, as it optimizes the continuous part of the space while exhaustively searching choices in the discrete subspace. You can customize this behavior to only sample a certain percentage of the discrete subspace via the ``sample_percentage`` attribute and to choose different sampling strategies via the ``hybrid_sampler`` attribute. An example on using this recommender in a hybrid space can be found [here](./../../examples/Backtesting/hybrid).
 
@@ -19,13 +19,13 @@ The [`NaiveHybridSpaceRecommender`](baybe.recommenders.naive.NaiveHybridRecommen
 ## Clustering recommenders
 
 BayBE offers a set of recommenders leveraging clustering techniques to facilitate initial point selection:
-* **[`PAMClusteringRecommender`](baybe.recommenders.nonpredictive.clustering.PAMClusteringRecommender):** This recommender utilizes partitioning around medoids for effective clustering.
-* **[`KMeansClusteringRecommender`](baybe.recommenders.nonpredictive.clustering.KMeansClusteringRecommender):** This recommender implements the k-means clustering strategy.
-* **[`GaussianMixtureClusteringRecommender`](baybe.recommenders.nonpredictive.clustering.GaussianMixtureClusteringRecommender):** This recommender leverages Gaussian Mixture Models for clustering.
+* **[`PAMClusteringRecommender`](baybe.recommenders.pure.nonpredictive.clustering.PAMClusteringRecommender):** This recommender utilizes partitioning around medoids for effective clustering.
+* **[`KMeansClusteringRecommender`](baybe.recommenders.pure.nonpredictive.clustering.KMeansClusteringRecommender):** This recommender implements the k-means clustering strategy.
+* **[`GaussianMixtureClusteringRecommender`](baybe.recommenders.pure.nonpredictive.clustering.GaussianMixtureClusteringRecommender):** This recommender leverages Gaussian Mixture Models for clustering.
 
 ## Sampling recommenders
 
 BayBE provides two sampling-based recommenders:
 
-* **[`RandomRecommender`](baybe.recommenders.nonpredictive.sampling.RandomRecommender):** This recommender offers random recommendations for all types of search spaces. This recommender is extensively used in backtesting examples, providing a valuable comparison. For detailed usage examples, refer to the examples listed [here](./../../examples/Backtesting/Backtesting).
-* **[`FPSRecommender`](baybe.recommenders.nonpredictive.sampling.FPSRecommender):** This recommender is only applicable for discrete search spaces, and recommends points based on farthest point sampling. A practical application showcasing the usage of this recommender can be found [here](./../../examples/Custom_Surrogates/surrogate_params).
+* **[`RandomRecommender`](baybe.recommenders.pure.nonpredictive.sampling.RandomRecommender):** This recommender offers random recommendations for all types of search spaces. This recommender is extensively used in backtesting examples, providing a valuable comparison. For detailed usage examples, refer to the examples listed [here](./../../examples/Backtesting/Backtesting).
+* **[`FPSRecommender`](baybe.recommenders.pure.nonpredictive.sampling.FPSRecommender):** This recommender is only applicable for discrete search spaces, and recommends points based on farthest point sampling. A practical application showcasing the usage of this recommender can be found [here](./../../examples/Custom_Surrogates/surrogate_params).
